+++ src/refactoring/components/cart/CouponSelector.tsx
@@ -1,9 +1,12 @@
 import { Coupon } from "../../../types";
-import { useCart } from "../../hooks";
 
-const CouponSelector = ({ coupons }: { coupons: Coupon[] }) => {
-  const { applyCoupon, selectedCoupon } = useCart();
+interface Props {
+  coupons: Coupon[];
+  applyCoupon: (coupon: Coupon) => void;
+  selectedCoupon: Coupon | null;
+}
 
+const CouponSelector = ({ coupons, applyCoupon, selectedCoupon }: Props) => {
   return (
     <div className="mt-6 bg-white p-4 rounded shadow">
       <h2 className="text-2xl font-semibold mb-2">쿠폰 적용</h2>
